proactively create whisper files for new metrics
If the metric is not in the cache AND the file for storing that metric does not exist, create the file to allow querying for those metrics immediately. This is helpful in the case that there is a large backlog of writing new metrics to disk which can cause a significant delay before being able to query for those metrics.