create whisper files in a separate thread/service
When receiving metrics that are not currently in the cache, append a tuple with the metric name and database path, to a queue, which can be used in another thread/service to proactively create the whisper file. The new service continually tries to pop a tuple from the queue and create the whisper file. It is possible that both threads make it past the existence check for the file, but in that case, the writer assumes the file exists and attempts to update it anyway.